requirements.txt
google-generativeai
requests
beautifulsoup4
python-dotenv